feat(desktop): extract DesktopDriverCapability + DesktopChassis::build (ADR-0071 phase 3)
The desktop chassis main.rs collapses from 1073 lines to 28: read env
into DesktopEnv, ask DesktopChassis::build() to assemble the substrate
+ driver, run the BuiltChassis. The winit App struct, ApplicationHandler
impl, helper fns, and the WORKERS const all move to a new driver.rs
module wrapped in DesktopDriverCapability + DesktopDriverRunning.

DesktopChassis::build() lives in chassis.rs alongside the existing
control-plane handler. It stands up SubstrateBoot, registers the inline
render + camera sinks (kept on the legacy register_sink path through
phase 3; phase 4 promotes render to a Capability), adds the existing
ADR-0070 native capabilities (audio, io, net, log) on the
boot.add_capability path, connects to the hub if AETHER_HUB_URL is set,
then wraps everything in a DesktopDriverCapability and hands it to the
chassis_builder::Builder. DesktopChassis is a unit struct now — the
chassis instance is the BuiltChassis<DesktopChassis>.

Phase-2A revision: drop the over-cautious Send + 'static bound on
DriverCapability and DriverRunning, and the +Send bound on PassiveBoot
and DriverBoot closures. winit's EventLoop is !Send on macOS, so the
driver capability and its running can't be Send. The chassis_builder
runtime is single-threaded between Builder construction and
BuiltChassis::run, so neither needs to cross threads. DriverRunning
keeps `'static` (load-bearing for the Box<dyn DriverRunning> trait
object).

The Chassis trait keeps its run() method through the migration; phases
4-7 will collapse it once every chassis switches to the build(env)?.run()
path. DesktopChassis::run on the marker struct returns an error
directing callers to use BuiltChassis::run.

Refs ADR-0071.
